Refactor kotlinImpl
