+++ kotlinImpl/src/main/kotlin/chaos/unity/nengao/KtPosition.kt
@@ -1,0 +1,6 @@
+package chaos.unity.nengao
+
+import chaos.unity.nenggao.AbstractPosition
+
+@Suppress("unused")
+data class KtPosition(val line: Int, val pos: Int) : AbstractPosition(line, pos)
