--- kotlinImpl/src/main/kotlin/chaos/unity/nengao/kotlin/Label.kt
@@ -1,6 +1,0 @@
-package chaos.unity.nengao.kotlin
-
-import chaos.unity.nenggao.AbstractLabel
-import chaos.unity.nenggao.AbstractSpan
-
-data class Label(val span: AbstractSpan, val message: String) : AbstractLabel(span, message)
