--- kotlinImpl/src/main/kotlin/chaos/unity/nengao/kotlin/Position.kt
@@ -1,5 +1,0 @@
-package chaos.unity.nengao.kotlin
-
-import chaos.unity.nenggao.AbstractPosition
-
-data class Position(val line: Int, val pos: Int) : AbstractPosition(line, pos)
