--- src/test/resources/long_line.yk
@@ -1,10 +1,0 @@
-start
-post-start
-
-
-
-
-
-
-pre-end
-end
